requirements.txt
alembic==1.5.8
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2==2.8.6
PyJWT==2.0.1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.16.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.4.4
Werkzeug==1.0.1
